Class Triangle: Added : Fields : side1, side2, side3 Constructors : edited all to include all sides Methods : Getters and Setters for all sides. perimeter, area and setSides methods
Fields:
For all three sides a field that is of type double was added

Constructors:
all Constructors initiated the sides to be 0

Methods:
Usual getters and setters. The perimeter method takes in all three sides as arguments and one method to set all sides that also takes in three arguments.

Other methods that don't take in any arguments, for the perimeter and area, these two methods assume that the sides have been assigned to the correct values.