Class Main:
All remaining 2D shapes : calRectangle, calSquare
Improved robustness with the allowing of double values in the dimensions such as radius, sides and height.

Improved readability for 2DObjects:
Created 2 boolean variables for the selection methods (calcArea and calcPerimeter). This improves the readability of the code so that one sees what the code is checking as a condition into the conditional statements.